Expression evaluator is now functional
Changed static class member to static methods to fix static initialization order
Changed type system to use dynamic_cast

+++ slib/collections/PriorityQueue.h
@@ -82,8 +82,8 @@ public:
 		return PRIORITYQUEUECLASS();
 	}
 
-	virtual ssize_t size() const override {
-		return (ssize_t)_size;
+	virtual size_t size() const override {
+		return _size;
 	}
 
 	using AbstractQueue<E>::isEmpty;
